Added Tessellator.getRenderPos(x, y, z) and changed Tessellator.drawText
Tessellator.drawText(text, x, y, z, color = white, renderBlackBox = true, scale = f, increase = true)
- Changed order of arguments to make it more usable
- Made some arguments have defaults

Removed all calls to javax.vecmath.Vector3d and javax.vecmath.Vector3f in favor of org.lwjgl.util.vector.Vector3f to be more consistant

put Image.getBufferedImage into a companion object so it doesn't get re-instanced every time a new image is created

+++ src/main/java/com/chattriggers/ctjs/minecraft/mixins/MixinRenderManager.java
@@ -3,12 +3,11 @@ package com.chattriggers.ctjs.minecraft.mixins;
 import com.chattriggers.ctjs.minecraft.wrappers.objects.Entity;
 import com.chattriggers.ctjs.triggers.TriggerType;
 import net.minecraft.client.renderer.entity.RenderManager;
+import org.lwjgl.util.vector.Vector3f;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
-
-import javax.vecmath.Vector3d;
 
 @Mixin({RenderManager.class})
 public class MixinRenderManager {
@@ -20,7 +19,7 @@ public class MixinRenderManager {
     private void onRenderEntity(net.minecraft.entity.Entity entity, double x, double y, double z, float entityYaw, float partialTicks, boolean hideDebugBox, CallbackInfoReturnable ci) {
         TriggerType.RENDER_ENTITY.triggerAll(
                 new Entity(entity),
-                new Vector3d(x, y, z),
+                new Vector3f((float) x, (float) y, (float) z),
                 partialTicks,
                 ci
         );
